[infra] move new certs to separate layer, extract common source of truth for custom domains [skip pizza]
