+++ infrastructure/common/types.ts
@@ -1,4 +1,37 @@
-export interface dbUrlArgs {
+// Secrets to be added to Pulumi when a new team is on-boarded
+export interface TeamSecrets {
+  // Token for the GovUKPayment service
+  // Provider by team
+  govUKPayToken: string;
+  // Uniform client details in the format clientId:clientSecret
+  // Each customer may have multiple clients (e.g. historic Uniform instances prior to council mergers)
+  // Provider by Idox
+  uniformClient?: string;
+};
+
+export interface Team {
+  // Must match "team_slug" in the PlanX "teams" table
+  name: string;
+  // Must match "name" in the Digital Land dataset "local-authority-district" (https://www.planning.data.gov.uk/dataset/local-authority-district)
+  uniformInstances?: string[];
+  govPayStagingOnly?: boolean;
+};
+
+export interface CustomDomain {
+  name: string,
+  domain: string,
+  certificateLocation?: "secretsManager" | "pulumiConfig",
+  // whether domain is served by legacy per-domain CloudFront distribution setup (defaults to false)
+  isLegacy?: boolean,
+  /**
+   * Whether the council has added DNS records for the shared ACM certificate.
+   * When false (default), the domain is added to the "mining" cert to surface validation records.
+   * When true, the domain is added to the shared cert as a SAN, and to the shared CDN as an alias.
+   */
+  isReady?: boolean,
+}
+
+export interface DbUrlArgs {
   role: string;
   password: string;
   host: string;
